+++ examples/noise_handshake.rs
@@ -1,12 +1,10 @@
 mod common;
-use common::{prefix_with_len, read_len_prefixed_message};
 
-use bytes::{Buf, BufMut};
-use parking_lot::{Mutex, RwLock};
-use tokio::{
-    io::{AsyncReadExt, AsyncWriteExt},
-    time::sleep,
-};
+use bytes::{Buf, BytesMut};
+use futures_util::{sink::SinkExt, TryStreamExt};
+use parking_lot::RwLock;
+use tokio::time::sleep;
+use tokio_util::codec::{Decoder, Encoder, Framed, FramedParts, LengthDelimitedCodec};
 use tracing::*;
 use tracing_subscriber::filter::LevelFilter;
 
@@ -18,17 +16,12 @@ use pea2pea::{
 use std::{collections::HashMap, io, net::SocketAddr, str, sync::Arc, time::Duration};
 
 // maximum noise message size, as specified by its protocol
-const NOISE_BUF_LEN: usize = 65535;
-
-struct NoiseState {
-    state: snow::TransportState,
-    buffer: Box<[u8]>, // an encryption/decryption buffer
-}
+const NOISE_MAX_LEN: usize = 65535;
 
 #[derive(Clone)]
 struct SecureNode {
     node: Node,
-    noise_states: Arc<RwLock<HashMap<SocketAddr, Arc<Mutex<NoiseState>>>>>,
+    noise_states: Arc<RwLock<HashMap<SocketAddr, NoiseState>>>,
 }
 
 impl Pea2Pea for SecureNode {
@@ -37,12 +30,120 @@ impl Pea2Pea for SecureNode {
     }
 }
 
+enum NoiseState {
+    Handshake(Box<snow::HandshakeState>),
+    PostHandshake(Arc<snow::StatelessTransportState>),
+}
+
+impl Clone for NoiseState {
+    fn clone(&self) -> Self {
+        Self::PostHandshake(Arc::clone(self.post_handshake()))
+    }
+}
+
+impl NoiseState {
+    fn handshake(&mut self) -> Option<&mut snow::HandshakeState> {
+        if let Self::Handshake(state) = self {
+            Some(state)
+        } else {
+            None
+        }
+    }
+
+    fn into_post_handshake(self) -> Self {
+        if let Self::Handshake(state) = self {
+            let state = state.into_stateless_transport_mode().unwrap();
+            Self::PostHandshake(Arc::new(state))
+        } else {
+            panic!();
+        }
+    }
+
+    fn post_handshake(&self) -> &Arc<snow::StatelessTransportState> {
+        if let Self::PostHandshake(state) = self {
+            state
+        } else {
+            panic!();
+        }
+    }
+}
+
+struct NoiseCodec {
+    codec: LengthDelimitedCodec,
+    noise: NoiseState,
+    buffer: Box<[u8]>,
+    nonce: u64,
+}
+
+impl NoiseCodec {
+    fn new(noise: NoiseState) -> Self {
+        NoiseCodec {
+            codec: LengthDelimitedCodec::builder()
+                .max_frame_length(NOISE_MAX_LEN)
+                .new_codec(),
+            noise,
+            nonce: 0,
+            buffer: vec![0u8; NOISE_MAX_LEN].into(),
+        }
+    }
+}
+
+impl Decoder for NoiseCodec {
+    type Item = String;
+    type Error = io::Error;
+
+    fn decode(&mut self, src: &mut BytesMut) -> Result<Option<Self::Item>, Self::Error> {
+        let bytes = if let Some(bytes) = self.codec.decode(src)? {
+            bytes
+        } else {
+            return Ok(None);
+        };
+
+        let msg_len = if let Some(noise) = self.noise.handshake() {
+            noise
+                .read_message(bytes.chunk(), &mut self.buffer)
+                .map_err(|_| io::ErrorKind::InvalidData)?
+        } else {
+            let len = self
+                .noise
+                .post_handshake()
+                .read_message(self.nonce, bytes.chunk(), &mut self.buffer)
+                .map_err(|_| io::ErrorKind::InvalidData)?;
+            self.nonce += 1;
+            len
+        };
+        let msg = String::from_utf8(self.buffer[..msg_len].to_vec()).unwrap();
+
+        Ok(Some(msg))
+    }
+}
+
+impl Encoder<String> for NoiseCodec {
+    type Error = io::Error;
+
+    fn encode(&mut self, msg: String, dst: &mut BytesMut) -> Result<(), Self::Error> {
+        let msg_len = if let Some(noise) = self.noise.handshake() {
+            noise.write_message(&[], &mut self.buffer).unwrap()
+        } else {
+            let len = self
+                .noise
+                .post_handshake()
+                .write_message(self.nonce, msg.as_bytes(), &mut self.buffer)
+                .map_err(|_| io::ErrorKind::InvalidData)?;
+            self.nonce += 1;
+            len
+        };
+        let msg = self.buffer[..msg_len].to_vec().into();
+
+        self.codec.encode(msg, dst)
+    }
+}
+
 impl SecureNode {
     // create a SecureNode
     async fn new(name: &str) -> io::Result<Self> {
         let config = Config {
             name: Some(name.into()),
-            read_buffer_size: NOISE_BUF_LEN + 2, // 2 for the encrypted message length,
             ..Default::default()
         };
         let node = Node::new(Some(config)).await?;
@@ -66,71 +167,57 @@ impl Handshake for SecureNode {
         let noise_builder = builder
             .local_private_key(&static_key)
             .psk(3, PRE_SHARED_KEY);
-        let mut buffer: Box<[u8]> = vec![0u8; NOISE_BUF_LEN].into();
-        let mut buf = [0u8; NOISE_BUF_LEN]; // a temporary intermediate buffer to decrypt from
 
-        let state = match !conn.side {
+        let stream = conn.take_stream();
+
+        let (stream, noise) = match !conn.side {
             ConnectionSide::Initiator => {
-                let mut noise = noise_builder.build_initiator().unwrap();
+                let noise = Box::new(noise_builder.build_initiator().unwrap());
+                let mut framed = Framed::new(stream, NoiseCodec::new(NoiseState::Handshake(noise)));
 
                 // -> e
-                let len = noise.write_message(&[], &mut buffer).unwrap();
-                conn.writer()
-                    .write_all(&prefix_with_len(2, &buffer[..len]))
-                    .await?;
+                framed.send("".into()).await?;
                 debug!(parent: self.node().span(), "sent e (XX handshake part 1/3)");
 
                 // <- e, ee, s, es
-                conn.reader().read(&mut buf).await?;
-                let message =
-                    read_len_prefixed_message::<_, 2>(&mut io::Cursor::new(buf))?.unwrap();
-                noise.read_message(&message, &mut buffer).unwrap();
+                framed.try_next().await?;
                 debug!(parent: self.node().span(), "received e, ee, s, es (XX handshake part 2/3)");
 
                 // -> s, se, psk
-                let len = noise.write_message(&[], &mut buffer).unwrap();
-                conn.writer()
-                    .write_all(&prefix_with_len(2, &buffer[..len]))
-                    .await?;
+                framed.send("".into()).await?;
                 debug!(parent: self.node().span(), "sent s, se, psk (XX handshake part 3/3)");
 
-                noise.into_transport_mode().unwrap()
+                let FramedParts { io, codec, .. } = framed.into_parts();
+                let NoiseCodec { noise, .. } = codec;
+                (io, noise.into_post_handshake())
             }
             ConnectionSide::Responder => {
-                let mut noise = noise_builder.build_responder().unwrap();
+                let noise = Box::new(noise_builder.build_responder().unwrap());
+                let mut framed = Framed::new(stream, NoiseCodec::new(NoiseState::Handshake(noise)));
 
                 // <- e
-                conn.reader().read(&mut buf).await?;
-                let message =
-                    read_len_prefixed_message::<_, 2>(&mut io::Cursor::new(buf))?.unwrap();
-                noise.read_message(&message, &mut buffer).unwrap();
+                framed.try_next().await?;
                 debug!(parent: self.node().span(), "received e (XX handshake part 1/3)");
 
                 // -> e, ee, s, es
-                let len = noise.write_message(&[], &mut buffer).unwrap();
-                conn.writer()
-                    .write_all(&prefix_with_len(2, &buffer[..len]))
-                    .await?;
+                framed.send("".into()).await?;
                 debug!(parent: self.node().span(), "sent e, ee, s, es (XX handshake part 2/3)");
 
                 // <- s, se, psk
-                conn.reader().read(&mut buf).await?;
-                let message =
-                    read_len_prefixed_message::<_, 2>(&mut io::Cursor::new(buf))?.unwrap();
-                noise.read_message(&message, &mut buffer).unwrap();
+                framed.try_next().await?;
                 debug!(parent: self.node().span(), "received s, se, psk (XX handshake part 3/3)");
 
-                noise.into_transport_mode().unwrap()
+                let FramedParts { io, codec, .. } = framed.into_parts();
+                let NoiseCodec { noise, .. } = codec;
+                (io, noise.into_post_handshake())
             }
         };
 
+        conn.return_stream(stream);
+
         debug!(parent: self.node().span(), "XX handshake complete");
 
-        let noise_state = NoiseState { state, buffer };
-
-        self.noise_states
-            .write()
-            .insert(conn.addr, Arc::new(Mutex::new(noise_state)));
+        self.noise_states.write().insert(conn.addr, noise);
 
         Ok(conn)
     }
@@ -139,25 +226,11 @@ impl Handshake for SecureNode {
 #[async_trait::async_trait]
 impl Reading for SecureNode {
     type Message = String;
+    type Codec = NoiseCodec;
 
-    fn read_message<R: Buf>(
-        &self,
-        source: SocketAddr,
-        reader: &mut R,
-    ) -> io::Result<Option<Self::Message>> {
-        let bytes = read_len_prefixed_message::<_, 2>(reader)?;
-
-        if let Some(bytes) = bytes {
-            let noise = Arc::clone(self.noise_states.read().get(&source).unwrap());
-            let NoiseState { state, buffer } = &mut *noise.lock();
-
-            let len = state.read_message(&bytes, buffer).ok().unwrap();
-            let decrypted_message = String::from_utf8(buffer[..len].to_vec()).unwrap();
-
-            Ok(Some(decrypted_message))
-        } else {
-            Ok(None)
-        }
+    fn codec(&self, addr: SocketAddr) -> Self::Codec {
+        let state = self.noise_states.read().get(&addr).cloned().unwrap();
+        NoiseCodec::new(state)
     }
 
     async fn process_message(&self, source: SocketAddr, message: Self::Message) -> io::Result<()> {
@@ -169,18 +242,11 @@ impl Reading for SecureNode {
 
 impl Writing for SecureNode {
     type Message = String;
+    type Codec = NoiseCodec;
 
-    fn write_message<B: BufMut>(&self, target: SocketAddr, payload: &Self::Message, buf: &mut B) {
-        info!(parent: self.node.span(), "sending an encrypted message to {}: \"{}\"", target, payload);
-
-        let noise = Arc::clone(self.noise_states.read().get(&target).unwrap());
-
-        let NoiseState { state, buffer } = &mut *noise.lock();
-        let len = state.write_message(payload.as_bytes(), buffer).unwrap();
-        let encrypted_message = &buffer[..len];
-
-        buf.put_u16_le(encrypted_message.len() as u16);
-        buf.put(encrypted_message);
+    fn codec(&self, addr: SocketAddr) -> Self::Codec {
+        let state = self.noise_states.write().remove(&addr).unwrap();
+        NoiseCodec::new(state)
     }
 }
 
@@ -204,25 +270,28 @@ async fn main() {
         .await
         .unwrap();
 
-    // allow a bit of time for the handshake process to conclude
+    // determine the initiator's address first
     sleep(Duration::from_millis(10)).await;
-
-    // send a message from initiator to responder
-    let msg = "why hello there, fellow noise protocol user; I'm the initiator";
-    initiator
-        .send_direct_message(responder.node().listening_addr().unwrap(), msg.to_string())
-        .unwrap()
-        .await
-        .unwrap();
-
-    // send a message from responder to initiator; determine the latter's address first
     let initiator_addr = responder.node().connected_addrs()[0];
-    let msg = "why hello there, fellow noise protocol user; I'm the responder";
-    responder
-        .send_direct_message(initiator_addr, msg.to_string())
-        .unwrap()
-        .await
-        .unwrap();
+
+    // send multiple messages to double-check nonce handling
+    for _ in 0..3 {
+        // send a message from initiator to responder
+        let msg = "why hello there, fellow noise protocol user; I'm the initiator";
+        initiator
+            .send_direct_message(responder.node().listening_addr().unwrap(), msg.to_string())
+            .unwrap()
+            .await
+            .unwrap();
+
+        // send a message from responder to initiator
+        let msg = "why hello there, fellow noise protocol user; I'm the responder";
+        responder
+            .send_direct_message(initiator_addr, msg.to_string())
+            .unwrap()
+            .await
+            .unwrap();
+    }
 
     sleep(Duration::from_millis(10)).await;
 }